+++ src/CellUnits.js
@@ -1,0 +1,6 @@
+const CellUnits = {
+    Day: 0,
+    Hour: 1
+};
+
+export default CellUnits