+++ src/DnDSource.js
@@ -30,8 +30,13 @@ export default class DnDSource {
 
                 if(isEvent) {
                     const event = item;
-                    if(viewType !== ViewTypes.Day) {
+                    if(config.relativeMove) {
                         newStart = localeMoment(event.start).add(localeMoment(newStart).diff(localeMoment(initialStart)), 'ms').format(DATETIME_FORMAT);
+                    } else {
+                        if(viewType !== ViewTypes.Day) {
+                            let tmpMoment = localeMoment(newStart);
+                            newStart = localeMoment(event.start).year(tmpMoment.year()).month(tmpMoment.month()).date(tmpMoment.date()).format(DATETIME_FORMAT);
+                        }
                     }
                     newEnd = localeMoment(newStart).add(localeMoment(event.end).diff(localeMoment(event.start)), 'ms').format(DATETIME_FORMAT);
                 }